chore: create reusage view for account address and connection
